requirements.txt
alembic==1.0.0
appdirs==1.4.3
appnope==0.1.0
arrow==0.12.1
atomicwrites==1.1.5
attrs==18.1.0
backcall==0.1.0
bcrypt==3.1.4
black==18.6b4
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
click==6.7
coverage==4.5.1
decorator==4.3.0
docker==3.5.0
docker-pycreds==0.3.0
factory-boy==2.11.1
Faker==0.9.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.12.1
Flask-Migrate==2.2.1
Flask-REST-JSONAPI==0.22.0
Flask-SQLAlchemy==2.3.2
idna==2.5
inflection==0.3.1
ipython==6.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.13.1
marshmallow-jsonapi==0.20.2
more-itertools==4.3.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
pluggy==0.7.1
prompt-toolkit==1.0.15
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
py==1.5.4
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.4
pytest==3.7.1
pytest-cov==2.5.1
pytest-dotenv==0.2.0
pytest-factoryboy==2.0.1
pytest-flask==0.10.0
pytest-mock==1.10.0
pytest-pspec==0.0.3
python-consul==1.1.0
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
python-nomad==1.0.0
redis==2.10.6
remote-pdb==1.2.0
requests==2.17.0
rq==0.12.0
rq-dashboard==0.3.12
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.10
text-unidecode==1.2
toml==0.9.4
traitlets==4.3.2
urllib3==1.21.1
wcwidth==0.1.7
websocket-client==0.51.0
Werkzeug==0.14.1
